refactor: EventSection to handle different views
